Temporary workaround for LP: #1055967
A security update in libdbus breaks ubiquity's kde frontend. Until the
"right" solution (converting this ubiquity frontend from kde to qt) is
officially added to precise, hold this security update back.

This doesn't really affect users much since they are getting the update
as part of the install anyway, so it will be included in the final system.